Fixed failed tests in PlayResumePlaybackEndpointTest. Renamed it from PlayResumeEndpointPlayerStateControllerTest.java. Added test configuration for the test suite
